Remove unused component
